[feat] Implementando página de cadastro de projeto e swap

+++ backend/src/main/java/ufrn/imd/cardeasy/models/CardList.java
@@ -26,7 +26,7 @@ public class CardList {
   private Integer id;
 
   @Column(nullable = false)
-  private Integer index;
+  private Long index;
 
   @Column(nullable = false)
   private String title;
